Separating System.exit from SpringApplication
* Moved CommandLineRunner to separate class SchemaSpyRunner.
* Removed the interface CommandLineRunner from SchemaSpyRunner.
* Added ExitCodeGenerator interface to SchemaSpyRunner.
* Moved CommandLine parse to SchemaSpyRunner.
* Main now create SpringApplication, retrieves SchemaSpyRunner,
  calls run on SchemaSpyRunner, exits the SpringApplication and
  calls System.exit
